Move event views
