+++ frontend/src/routes.js
@@ -4,8 +4,8 @@ import { syncHistoryWithStore } from 'react-router-redux'
 
 import store from 'store'
 import Layout from 'views/layout'
-import EventsSearch from 'views/events-search'
-import EventDetails from 'views/event-details'
+import EventsSearch from 'views/events/search'
+import EventDetails from 'views/events/details'
 import Dashboard from 'views/dashboard'
 import FailuresSearch from 'views/failures/search'
 import FailureDetails from 'views/failures/details'
