+++ frontend/src/views/events/details/index.js
