+++ frontend/src/views/events/search/index.js
