+++ frontend/src/views/events/search/index.scss
